Initial PR feedback
- Moved hash functions to utils
- fixed typo
