+++ learning_observer/learning_observer/log_event.py
@@ -62,6 +62,7 @@ import learning_observer.filesystem_state
 import learning_observer.paths as paths
 import learning_observer.settings as settings
 import learning_observer.prestartup
+from learning_observer.util import secure_hash
 
 
 # These should move into the startup check
@@ -187,35 +188,6 @@ def encode_json_block(block):
     the same string, with the same hash.
     '''
     return json.dumps(block, sort_keys=True, indent=3)
-
-
-def secure_hash(text):
-    '''
-    Our standard hash functions. We can either use either
-
-    * A full hash (e.g. SHA3 512) which should be secure against
-    intentional attacks (e.g. a well-resourced entity wants to temper
-    with our data, or if Moore's Law starts up again, a well-resourced
-    teenager).
-
-    * A short hash (e.g. MD5), which is no longer considered
-    cryptographically-secure, but is good enough to deter casual
-    tempering. Most "tempering" comes from bugs, rather than attackers,
-    so this is very helpful still. MD5 hashes are a bit more manageable
-    in size.
-
-    For now, we're using full hashes everywhere, but it would probably
-    make sense to alternate as makes sense. MD5 is 32 characters, while
-    SHA3_512 is 128 characters (104 if we B32 encode).
-    '''
-    return "SHA512_" + hashlib.sha3_512(text).hexdigest()
-
-
-def insecure_hash(text):
-    '''
-    See `secure_hash` above for documentation
-    '''
-    return "MD5_" + hashlib.md5(text).hexdigest()
 
 
 def log_event(event, filename=None, preencoded=False, timestamp=False):
