+++ modules/writing_observer/writing_observer/awe_nlp.py
@@ -25,7 +25,7 @@ import time
 import warnings
 
 import writing_observer.nlp_indicators
-from learning_observer import kvs
+from learning_observer import kvs, util
 
 RUN_MODES = enum.Enum('RUN_MODES', 'MULTIPROCESSING SERIAL')
 
@@ -198,17 +198,9 @@ async def process_texts_parallel(texts, options=None):
     return annotated
 
 
-def hash_cache_key(text):
-    '''
-    Create cache key based on text
-    '''
-    hash_key = hashlib.sha1(text.encode('utf-8')).hexdigest()
-    return 'NLP_CACHE_' + hash_key
-
-
 async def process_texts(writing_data, options=None, mode=RUN_MODES.MULTIPROCESSING):
     '''
-    Process texts with cachinng
+    Process texts with caching
     1. Create hash of text
     2. Fetch cache data
     3. Check to see if all options are present
@@ -229,7 +221,9 @@ async def process_texts(writing_data, options=None, mode=RUN_MODES.MULTIPROCESSI
 
     for writing in writing_data:
         text = writing.get('text', '')
-        text_hash = hash_cache_key(text)
+        if len(text) == 0:
+            continue
+        text_hash = 'NLP_CACHE_' + util.secure_hash(text.encode('utf-8'))
         text_cache_data = await cache[text_hash]
         if text_cache_data is None:
             text_cache_data = {}
@@ -244,15 +238,11 @@ async def process_texts(writing_data, options=None, mode=RUN_MODES.MULTIPROCESSI
     if len(need_processing) > 0:
         just_the_text = [w.get("text", "") for w in need_processing]
         annotated_texts = await processor[mode](just_the_text, list(needed_options))
-        
+
         for annotated_text, single_doc in zip(annotated_texts, need_processing):
             if annotated_text != "Error":
                 single_doc.update(annotated_text)
-                text_hash = hash_cache_key(single_doc.get('text', ''))
-                # which is the lesser of two evils?
-                # we store the student key in the cache
-                # OR 
-                # we remove it before storing (uses dict comprehension)
+                text_hash = 'NLP_CACHE_' + util.secure_hash(single_doc['text'].encode('utf-8'))
                 new_cache = {k: v for k, v in single_doc.items() if k != 'student'}
                 await cache.set(text_hash, new_cache)
         results.extend(need_processing)
